Add ipv6 functionality
This opens a symmetric pair of ports on an ipv6 based port.  There is no attempt to stop two different devices from joining the same server but on ipv4 vs ipv6.

Both ipv4 and ipv6 will get the same bound callbacks, so the client of this will not have to care where its traffic originates